pyproject.toml
[project]
name = "BuesinessAgents"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]

[tool.setuptools]
packages = [
    "agent_openai",
    "llm_auditor",
    "multi_tool_agent",
    "my_agent",
    "travel_concierge",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"


[tool.ruff]
line-length = 80
target-version = "py313"

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehension
    "PL", # Pylint rules
    "B", # flake8-bugbear
    "UP", # pyupgrade
    "RUF" # ruff specific rules
]
ignore = [
    "E501", # Ignore line length for long LLM prompts
    "PLR0913" # Allow agents to have many arguments if needed
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"


